==> 00_setup.rmd <==
---
title: "00_setup.rmd"
output: github_document
---

## Run this once at start
## Make sure to set the working directory properly

```{r}
options(default.stringsAsFactors(FALSE))
set.seed(123)

dir.create("data", showWarnings = FALSE)
dir.create("data/raw", showWarnings = FALSE)
dir.create("data/processed", showWarnings = FALSE, recursive = TRUE)
dir.create("results", showWarnings = FALSE)
dir.create("results/figures", showWarnings = FALSE, recursive = TRUE)
dir.create("results/tables", showWarnings = FALSE, recursive = TRUE)
```

# packages

```{r }
pkgs <- c(
    "TCGAbiolinks",
    "dplyr",
    "stringr",
    "readr",
    "tibble"
)

to_install <- pkgs[!vapply(pkgs, requireNamespace, logical(1), quietly = TRUE)]

if (length(to_install) > 0){
    message("Installing missing packages: ", paste(to_install, collapse = ","))
    install.packages(to_install)
}

if(!requireNamespace("Biocmanager", quietly = TRUE)){
    install.packages("Biocmanager")
}

if(!requireNamespace("TCGAbiolinks", quietly = TRUE)){
    BiocManager::install("TCGAbiolinks", ask = FALSE, update = FALSE)
}

message("Setup Completed")
```
